requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics
scikit-learn>=1.3.0

# Progress bars
tqdm>=4.65.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Configuration (optional)
pyyaml>=6.0